requirements.txt
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0
cohere>=4.47
groq>=0.4.0
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
tenacity>=8.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
